order-service/requirements.txt
blinker==1.8.2
click==8.1.7
Flask==3.0.3
grpcio==1.64.0
grpcio-tools==1.64.0
importlib_metadata==7.1.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mysql-connector-python==8.4.0
protobuf==5.27.0
Werkzeug==3.0.3
zipp==3.19.0
